refactor: Add emoed method to ListComponent
This commit adds a new method called `emoed` to the `ListComponent` in the `feed/list` module. The `emoed` method is used to check if a post has been liked by the current user. It returns a boolean value based on whether the user's ID is found in the `likes` array of the post object.

The purpose of this change is to provide a way to determine if a post has been liked by the user, which can be used to display the appropriate UI elements.